moved spe.py file to scripts/
